pyproject.toml
[tool.poetry]
name = "food-calculator"
version = "0.1.0"
description = "Django pet project..."
authors = ["Your Name <you@example.com>"]
license = "BSD"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.0"
asgiref = "3.7.2"
Django = "3.2.19"
djangorestframework = "3.14.0"
et-xmlfile = "1.1.0"
greenlet = "2.0.2"
mysqlclient = "2.1.0"
numpy = "1.25.1"
openpyxl = "3.1.2"
pandas = "2.0.3"
Pillow = "10.0.0"
python-dateutil = "2.8.2"
python-decouple = "3.8"
pytz = "2023.3"
six = "1.16.0"
SQLAlchemy = "2.0.18"
sqlparse = "0.4.4"
typing_extensions = "4.7.1"
tzdata = "2023.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
